refactor: simplify activeCell modules
