+++ src/contentScript/__tests__/nestedEditorLifecycle.test.ts
@@ -19,8 +19,8 @@ import { exitSourceModeEffect, sourceModeField, toggleSourceModeEffect } from '.
 import { rebuildTableWidgetsEffect } from '../tableState/tableWidgetEffects';
 import { markdown } from '@codemirror/lang-markdown';
 import { GFM } from '@lezer/markdown';
-import { resolveActiveCell } from '../tableRuntime/activeCell/activeCellResolver';
-import { requestOpenActiveCellEffect } from '../tableRuntime/activeCell/activeCellOpen';
+import { resolveActiveCell } from '../tableRuntime/activeCell/resolvedActiveCell';
+import { requestOpenCellEffect } from '../tableRuntime/openCellRequest';
 import {
     beginOpenCellRequestEffect,
     getPendingOpenCellRequest,
@@ -61,7 +61,7 @@ function openRequestEffects(params: {
 
     return [
         beginOpenCellRequestEffect.of(request),
-        requestOpenActiveCellEffect.of({
+        requestOpenCellEffect.of({
             requestId: params.requestId,
         }),
     ];
@@ -473,7 +473,7 @@ describe('nestedEditorLifecycle', () => {
         });
 
         view.dispatch({
-            effects: requestOpenActiveCellEffect.of({ requestId: 'missing-request' }),
+            effects: requestOpenCellEffect.of({ requestId: 'missing-request' }),
         });
         flushAnimationFrames();
 
@@ -537,7 +537,7 @@ describe('nestedEditorLifecycle', () => {
                     ) &&
                     effects.some(
                         (effect) =>
-                            effect?.is?.(requestOpenActiveCellEffect) && effect.value?.requestId === 'request-normalize'
+                            effect?.is?.(requestOpenCellEffect) && effect.value?.requestId === 'request-normalize'
                     )
                 );
             });
